+++ HISTORY.rst
@@ -2,6 +2,23 @@
 History
 =======
 
+
+0.2.1 (2018)
+------------
+
+* Allow for using mdbtools also on win
+* Slightly faster find_files using cache and fnmatch
+* Bug fix: error in sorting files when using pathlib fixed
+
+
+0.2.0 (2018-10-17)
+------------------
+
+* Improved creation of step tables (much faster)
+* Default compression on cellpy (hdf5) files
+* Bug fixes
+
+
 0.1.22 (2018-07-17)
 -------------------